requirements.txt
fastapi
requests
uvicorn
datetime
boto3
python-dotenv